requirements.txt
Django==4.2.7
djangorestframework==3.14.0
PyJWT==2.8.0
Pillow==10.0.1
requests==2.31.0
geopy==2.3.0
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
django-filter==23.3
